# openapi.yaml
openapi: 3.0.0
info:
  title: BRDG JPMP API
  description: API for the BRDG JPMP kiosk application
  version: 1.0.0
servers:
  - url: https://api.jpmp.brdg.jarv.us
    description: Production backend
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /orders:
    get:
      summary: Get all orders
      description: Returns a list of all orders, optionally filtered by user and/or kiosk type
      parameters:
        - name: latest
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Optional limit on the number of most recent orders to return
        - name: user
          in: query
          required: false
          schema:
            type: string
          description: Optional user ID to filter orders by user_profile->>'id'
        - name: kiosk_type
          in: query
          required: false
          schema:
            type: string
            enum: [sweet, juice]
          description: Optional kiosk type to filter orders by their kiosk_type
      responses:
        "200":
          description: A list of orders
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
        "400":
          description: Invalid query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      summary: Create a new order
      description: Creates a new order. Optionally set status to 'completed' for immediate completion.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - kiosk_id
                - items
              properties:
                kiosk_id:
                  type: integer
                items:
                  type: array
                  items:
                    type: object
                    required:
                      - id
                    properties:
                      id:
                        oneOf:
                          - type: integer
                            description: Item ID
                          - type: string
                            description: Item slug
                      customizations:
                        type: object
                user_profile:
                  type: object
                  description: User profile data
                status:
                  type: string
                  enum: [completed]
                  description: Optional status to set on creation. Only 'completed' is allowed. If not provided, defaults to 'pending'.
      responses:
        "201":
          description: Order created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /orders/status/{status}:
    get:
      summary: Get orders by status
      description: Returns a list of orders with a specific status
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
            enum: [pending, completed, canceled]
        - name: latest
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
          description: Optional limit on the number of most recent orders to return
      responses:
        "200":
          description: A list of orders with the specified status
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items:
                      $ref: "#/components/schemas/Order"
        "400":
          description: Invalid status or query parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /orders/{id}:
    get:
      summary: Get a specific order
      description: Returns a specific order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The requested order
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update an order
      description: Updates an existing order's status and/or survey response
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [pending, completed, canceled]
                  description: Order status
                survey_response:
                  type: object
                  description: Survey response data as JSON object
                  nullable: true
      responses:
        "200":
          description: Order updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: "#/components/schemas/Order"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      summary: Delete an order
      description: Deletes an existing order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        kiosk_id:
          type: integer
        kiosk_type:
          type: string
          enum: [sweet, juice]
          description: The type of kiosk at the time the order was created. This value is captured when the order is created and does not change even if the kiosk's type changes later.
        status:
          type: string
          enum: [pending, completed, canceled]
        user_profile:
          type: object
          description: User profile data
        survey_response:
          type: object
          description: Survey response data as JSON object
          nullable: true
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        order_id:
          type: integer
        item_id:
          type: integer
        slug:
          type: string
          description: Unique identifier for the item
        name:
          type: string
        description:
          type: string
        available:
          type: boolean
          description: Whether the item is currently available for ordering
        created_at:
          type: string
          format: date-time
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            message:
              type: string
            status:
              type: integer